+++ src/main/java/com/zuxelus/energycontrol/renderers/TERemoteThermoRenderer.java
@@ -19,7 +19,7 @@ public class TERemoteThermoRenderer extends TileEntitySpecialRenderer<TileEntity
 	@Override
 	public void renderTileEntityAt(TileEntityRemoteThermo te, double x, double y, double z, float partialTicks, int destroyStage) {
 		GlStateManager.pushMatrix();
-		GlStateManager.translate((float)x, (float)y, (float)z);
+		GlStateManager.translate(x, y, z);
 		EnumFacing facing = te.getFacing();
 		switch (facing) {
 		case UP:
